requirement.txt
alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
DateTime==5.5
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.8.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.1.0
SQLAlchemy==2.0.35
SQLAlchemy-Utc==0.14.0
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
utcdatetime==0.0.7
Werkzeug==3.0.4
zope.interface==7.0.3
boto3==1.35.49
statsd==4.0.1
requests==2.32.3
